+++ distage/distage-app/src/test/scala/com/github/pshirshov/izumi/distage/app/StaticPluginCheckerTest.scala
@@ -1,6 +1,5 @@
 package com.github.pshirshov.izumi.distage.app
 
-import com.github.pshirshov.izumi.distage.app.macros.{LogstageModuleRequirements, StaticPluginChecker}
 import com.github.pshirshov.test.plugins.{DependingPlugin, StaticTestPlugin}
 import org.scalatest.WordSpec
